add policy fix type to big decimal for number value implement interest rate service

+++ grails-app/domain/th/co/opendream/cashcard/domain/InterestRate.groovy
@@ -3,13 +3,13 @@ package th.co.opendream.cashcard.domain
 class InterestRate {
     Date startDate
     Date endDate
-    Float rate
+    BigDecimal rate
     Date dateCreated
     Date lastUpdated
 
     static constraints = {
         startDate unique: true, nullable: false, blank: false
         endDate unique: true, nullable: false, blank: false, validator: { val, obj -> val >= obj.startDate?.plus(1) }
-        rate min: 0F, nullable: false, blank: false
+        rate min: 0.00, nullable: false, blank: false
     }
 }
